requirements.txt
Django>=1.11,<1.12
configobj       # required by brewpi.py

pytz            # required by manage.py migrate
raven           # OPTIONAL - used for debugging
django-constance[database]  # for managing user-configured constants
GitPython                   # for managing git integration
zeroconf        # for locating ESP8266 devices
circus          # for managing Fermentrack wsgi & brewpi.py processes
#circus-web      # for managing Fermentrack wsgi & brewpi.py processes
chaussette      # for managing Fermentrack wsgi
huey>=2.0       # asynchronous task queue
pid             # for pid-file handling
pyserial        # for serial interaction

pyudev          # for managing udev rules for serial devices
requests        # for loading firmware data from websites

esptool         # for flashing ESP8266 devices

redis           # for huey & gravity sensor support

pybluez         # for gravity sensor support
# beacontools[scan]           # for Tilt hydrometer support
aioblescan      # Replacement for beacontools for Tilt support

# Requirements to make Circus work
pyzmq<17
tornado<5

# Django Rest Framework and related
djangorestframework
djangorestframework-jwt
python-dateutil
